Refactor internal image loading APIs
